Set an unusable password on new User records
We inherit from AbstractBaseUser to implement our custom User class.
This comes with some conveniences (eg - a last_login field), but also
bundles a password on the record.

We authenticate with CIS2 and don't use local passwords. We could
replace AbstractBaseUser and write a custom base, but a simpler change
is to set the password to something unusable.

Along with this, we override get_session_auth_hash, a function that
invalidates the session hash on password changes. Given we don't have
the concept of password changes in our app, here we simply hash the
user's is_active flag instead. This has the useful side effect of being
able to invalidate user sessions by toggling this flag.