use inline radios for right/left breast
The design system has an option for inline radios, but it does not
support conditionals, so we have to implement something ourselves.

Instead of rendering the radios component once, we will render it 3
times, once for each option, and use grid classes to align the options
how we want.

This breaks the support for error handling, so we need to reimplement
that outside of the component, by:

1. wrapping the fieldset in another div, so that the red line
   is shown next to the field
2. rendering the error message ourselves
3. not passing params.errorMessage to any of the radios calls

There is a minor inconsistency with how other fields show their errors
at the moment, this is due to another bug, causing us to render
the legends for the other fields outside of the div with the red line.
This can be addressed separately in a future commit.

The taborder is the same as with stacked radios, i.e. "Right" -> "Left"
-> "Other". Arrow navigation also follows this order even though "Other"
may appear directly to the right of "Right".

The labels and legend are unchanged and work as before. If a radio is
not selected, clicking on "Select the location of the lump" focuses
the "Right" radio, as before.

For now I've implemented this as a partial, rather than making a
reusable django form field/widget for this. If needed we can refactor
this when we have more than one form using it.