add notification banners to base layout
The message framework doesn't have a way to pass structured data,
but we can pass marked-safe HTML, as in this example from stack
overflow[1].

This is complicated by us using Jinja, which has it's own way of
marking strings safe, which is not recognised by the messages
framework. So when *adding* an HTML message, we need to mark it
with `mark_safe`, and when *rendering* the message we need to
use `Markup`.

Supporting html isn't essential, but it means we can show arbitrary
content in banners, including the success banner with header shown
in the design system example[2]

[1]: https://stackoverflow.com/questions/2053258/how-do-i-output-html-in-a-message-in-the-new-django-messages-framework
[2]: https://service-manual.nhs.uk/design-system/components/notification-banners