Add repositories for Appointments, Participants and Clinics
When retrieving these records, we need to make sure that database
queries are scoped to the current_provider set in the user session. This
helps ensure that users can't access records that don't belong to the
selected provider, even if their account technically has access via
their UserAssignments.

To help with this, we introduce repository objects for these three
models. The expectation is that all views will use these repositories
rather than interact with the models directly.

These repositories inherit from the ProviderScopedRepository class,
ensuring that a provider is given for queries to be scoped to.

Each repository implements common functions for filtering, ordering and
prefetching/selecting related records.

A linter to detect imports of these models in the view layer will be
added in a later commit.