Add ProviderScopedRepository base class for provider-scoped data access
Some of the data access in this service needs be scoped to the
current_provider set in the user session. To help with this, we add a
base repository class that enforces provider-scoped queries.

Subclasses specify the model and implement _scoped_queryset() to filter
records by provider and set this query on the _queryset attributes.

Subclasses can also expose methods that add additional query methods to
_queryset. These methods can return self in order to make them
chainable.

A final call to all() (implemented by this base class) will return the
final queryset.