refactor custom event method for App Insights
adding fields under the `extra` attribute ensures
an event is triggered in Insights[1]. Having an
event allows us to easily surface the stack trace
and transactions and alert on something, but the
message sent to the logger does not seem to appear
in the reporting around the event on Insights - it
needs to be sent in the `additional_attrs` field.
As with the other App Insights logs, this will
appear in Insights *and* our stdout logs for the
container.

[1]https://learn.microsoft.com/en-us/azure/azure-monitor/app/opentelemetry-add-modify?tabs=python#send-custom-events