disable notification autofocus by default
As per the prototype, we don't want notifications autofocused,
particularly on appointment pages, as they are non-essential
and distract from the task.

We are assuming the default autofocus behaviour is helpful for screen
reader users, but in our case we know all radiologists taking the
mammograms will be sighted.

This can be overriden if we need to make exceptions for other pages.