Switch to docker-container driver and Buildx
As per the build-and-push-docker-images documentation[1], setup-buildx
is not required but is recommended.

In our case, it will allow us to use the docker-container driver[2]
instead of the default docker driver when building images, which allows
us to use "cache export" functionality for registry caching[3].

Registry caching supports `max` cache mode, which allows for caching
of layers in multi-stage builds that are not exposed in the final
image.

[1]. https://github.com/marketplace/actions/build-and-push-docker-images
[2]. https://docs.docker.com/build/builders/drivers/
[3]. https://docs.docker.com/build/ci/github-actions/cache/