Remove all usage of get_object_or_404
We want to write a linter to discourage the use of direct record
retrieval from Appointment, Clinic, and Participants models. The linter
will look for 'ModelName.objects'. With this in mind, it would be useful
to discourage use of get_object_or_404, which simply accepts the bare
ModelName as an argument.

We'll also likely replace get_object_or_404 at some point with some
custom behaviour.

Replace use of this method with a slightly more verbose alternative.