implement multiple caches
See https://docs.docker.com/build/cache/backends/#multiple-caches

Caching to and from :buildcache means branches will push a build
cache that overwrites the other.

A common pattern is to use the cache of both the current branch and the main branch,
for the `cache-from` of builds triggered by PRs.

On main, we will just use the main cache.