move current_provider attribute to the user object
The django `user.has_perm` API reserves the `obj` parameter for
object-level permissions, e.g. `user.has_perm('books:read', obj=the_hobbit)`

In our case, we've overloaded that parameter for the logged in user's current
provider. This means we can't implement any permissions that grant
access for a particular object.

If we instead make `current_provider` an attribute of `request.user`,
we can use obj for its intended purpose.

For example, in appointment views, the permissions object should be the
appointment. The actions a user can do on an appointment depend on its
current state. By passing that appointment as the object we can
include that logic in the predicates in rules.py modules.

I ruled out adding extra **kwargs to user.has_perm, because it breaks
the assumptions of PermissionRequiredMixin and the permission_required
decorator.

I considered creating an extra AppointmentPolicy class and using
that to handle the object-level permissions, but then we have
authorization logic split between rules.py and policy classes
which seems awkward and unnecessary.