add START_MAMMOGRAM_APPOINTMENT permission
Add a predicate and a permission for starting a mammogram appointment.
This is not used yet, but will be in an upcoming PR to add the "Start
this appointment" button and the IN_PROGRESS appointment state.

This is an example of an object permission where we pass in the obj
param to user.has_perm.

I've moved the mammograms specific rules into the app, since otherwise
the existing module would get very long and the tests hard to read.
Django rules collects all these rules.py files as part of its
initialisation.