Add an integration test to verify blob file ordering
The Azure storage blob SDK for python provides a list_blobs method on the container client. We may receive multiple files per day and it is important they are processed in the correct sequential order denoted by filename.
This commit adds a test to show that for an appointment booked in one file and cancelled in another, the file with the booking is processed before the file with the cancellation.