add service for managing appointment state
This introduces a bit of control over when appointments can transition
to different state.

If a user triggers the same action multiple times, it should
have the same effect as if they had triggered it once.

This can happen with the checkout button javascript, since if the
fetch() errors, it falls back to submitting the form normally. If the
initial fetch() succeeded in creating the record, then then we want the
subsequent form submission to reflect this.

A consequence of this is we shouldn't do the state checking in the rules
functions. In the presenters/templates we will be checking if the state
is in one of the valid from_states, but in the updater we will check if
it's one of those or it's in the to_state already.