update to latest LTS version of node
node 24 is now in active LTS.
(https://nodejs.org/en/about/previous-releases)

We should pin npm in addition to node to prevent dependabot from using a
different version to what we're using locally.

Dependabot uses a different version of node, which generates some
warnings, but we think this is fine, and have made a similar update
already in the nhsuk-frontend repo. We will also have some temporary
warnings from nhsuk-frontend while we wait for the next release.