Use date objects when creating reports
There was a lack of test coverage around how we called the report generation from pandas.read_sql and this meant we were passing datetime objects instead of dates. This meant the failure (invites not sent) and reconciliaton reports were always empty. The query tests passed date objects and this differed from the implementation