Make checkin component emit an event on success
This will allow us to wire up handlers for showing/hiding the start
appointment button, which will be a separate component.

I considered embedding this logic into the check-in itself, but
logically these are two separate components, and on the appointment
details page the start appointment button is presented differently.