exempt persona login from csrf protection
If you log in, and select a persona, it's all fine and you are logged in
as that persona.

But if you then use the browser back button, and choose a different
persona you get a CSRF error. This is because django rotates the CSRF
token after logging in, so the previous CSRF token is invalid.

The persona login is only ever used in non-production environments,
so we don't need the CSRF protection. Exempting it means the user
can hit back in their browser and the page will still work.