Amend Appointment#current_status implementation
Using order_by in this method busts the prefetch cache, causing n+1
queries on the clinic show page (where we list all appointments).

Change the approach here so we can eliminate the n+1 issue and still
find the most recent status for an appointment without resorting to
sorting in Python.

- Add a function `eager_load_current_status` to AppointmentQuerySet
  * This prefetches the most recent status for each appointment, and
    sets it on an attribute `_prefetched_current_status`
  * We also select the related User for each status, as determined by
    the `created_by` field. We'll need this later when we start
    displaying the name of the user who set the status.
- Update `current_status` to check `_prefetched_current_status`. If it
  doesn't find anything, retain the original behaviour of fetching or
  returning a default status.