display the history item on medical information page
This is a placeholder for now. We're not nearly handling cases where
there are no history items, or multiple, but they will show up.

I've removed the dummy link, because we will need one for each type of
medical information. We can deal with that when we have a form to link
to.