add a basic presenter for breast cancer history items
This is just a placeholder for now; we want to reflect the data
on the record medical information page, so we can start implementing
the forms.

Later we will need to deal with formatting information for the
right/left breast side to side, and allow for hiding optional
information that is not provided.