Add a Validator to pair with `data-checkbox-exclusive`
This provides server side validation for ArrayFields that have
predefined choices as their base_field, where one of those
choices excludes all the others. For example a "none of the above"
option.

On the frontend, we have javascript to automatically deselect the other
checkboxes when such options are selected, but we need to make sure the
backend prevents bad data reaching the database, for example if the user
disables javascript, or if we are bypassing UI and entering data
directly into the db.