start form for breast cancer history
There are two new patterns here for arranging checkboxes
and radios into a grid. This is because we want to show
options relating to the right breast on the left
and left breast on the right, to match the perspective
of the mammographer facing the participant.

In one case, we are treating right breast and left breast as separate
fields, and rendering them side by side. Both fields have their own
legends and are validated independently.

In the other case, we model the question as a single field and we just
arrange the options. We implement this by calling the component twice
for the same field, but passing in different options.