add django-linear-migrations
Out of the box, django's migration system creates a lot of workflow
conflict when multiple people are working on the same app.

This is because it allows you to create two conflicting migrations
on parallel branches, which both pass CI, but then if the branches
are merged, it wants you to create a merge migration to reconcile
the migrations after the fact. This situation is currently
causing our main branch, and anything branched off it, to fail the
tests.

django-linear-migrations solves this problem by introducing "max
migration" files. These have the effect of turning a migration
conflict into a git conflict. Even if a PR hasn't been rebased,
it will conflict with main as soon as the conflicting migration
is merged to main, which will force us to deal with it before the
second PR is merged.

django-linear-migrations also introduces a `rebase_migration`
command we can use as an alternative to `makemigrations --merge`.