Allow modified labels for screenreaders
There are many use cases where we would want to vary the label or legend
shown to screenreader users, by adding visually hidden text before or
after the main label.

This commit adds extra params to custom form fields so we can specify
this text in the same place we specify the labels normally, i.e. in the
form class. I've extended this to most of the common field types
(radios, checkboxes, textareas and inputs).