Update github actions concurrency
In some occasions the delete review app workflow was cancelled as it
conflicted with the main deploy workflow. This is because the GITHUB_REF
on the delete workflow is actually main, as opposed to the PR branch.

Remove the branch reference which is prone to error and use the PR
number instead as it is more deterministic.