Main workflow concurrency
When the concurrency is set at the deploy job level, it is decoupled
from the tests. There is a chance a workflow run with slow tests gets
overtaken by a workflow run with fast tests which then deploys first. And
the first workflow run waits, then deploys later with an older commit.
This has been observed already.

Setting concurrency at the workflow level ensures a workflow run waits
the active run to finish completely before starting.