add a helper to generate HTML for conditional fields
The design system radios and checkboxes components take a "conditional"
param when defining the options. This allows you to add arbitrary html
that is conditionally shown when the option is selected.

Now that we have FormWithConditionalFields, we can declare the
relationships between conditionally dependent fields.

This commit extends that so that the conditional param is generated
based on those declarations, so you don't need to explicitly tell
the field what the conditional html is in the template.

The old method for setting arbitrary html is still there in case we need
it for more complex cases, where we might want to conditionally show
more than just a simple field.