Do not download report at end of smoke test
There is a potential security issue with the ADO pipeline having permissions to download a report from blob storage.
There are alternatives to this approach like searching logs which can be used to verify the smoke test.
As an interim step, do not download the smoke test report, simply verify that all container app jobs ran sucessfully.