Export ENVIRONMENT prior to running smoke test
The review environment exports ENVIRONMENT and the smoke test defaults to dev when looking up this environment variable. So the smoke test passes for these environments but fails on preprod as the variable is not set. Ensure we export it prior to running the test.

Removes some logic around PR_NUMBER and storage account details as these
aren't required anymore.