Benign lumps: make YearField bounds callable
The current YearField validation on min and max value is implemented in
a way that means values based on the current year will only be evaluated
on module import. An application restart after moving from one year to
the next would be needed to ensure current year is correct.

This isn't a huge bug necessarily with our current use cases, but it's
an inconvenient issue to leave in the codebase.

This commit switches to using min_value_callable and max_value_callable
to determine the year bounds. These can be overridden, but by default
specify a range of 80 years to current year.