add cooldown for application dependency updates
Dependabot's cooldown function is a protection against supply chain
attacks. There have been a lot of these recently against npm.

By waiting a week before pulling new package versions, we are much
less likely to be exposed to malicious versions before they are yanked
from the package index.

We may wish to tweak this for particular packages, e.g. if they are
internally authored. We are already excluding nhsuk-frontend from
dependabot so we don't need to worry about that one.

I'm not sure about terraform updates so I've left that alone in this
commit.