refactor: extract generic view for deletion confirmation
We already have a generic template for these pages, and I expect each to
follow a common pattern, so this is a case where it might be useful to use
Django's generic views.

We also need to make sure that all deletes are audited, and that we show
a success message.