change add links to small buttons
This leaves some empty summary list rows in the medical info summary
list, but this is only a placeholder that we will soon be replacing
with the proper design.

We've added horizontal rules between the summary lists and button groups
because there have been updates to the design system to remove extra
space below summary lists within cards, and these are
conceptually different so need separating a bit.