make participant:show redirect to mammograms:participant_details
We've removed the separate section for participants, so the only way to
view a participant now is in the context of an appointment.

However, it still might be useful to preserve the /participants/<pk> URL
space, as we have some data that is scoped to participant instead
of an appointment, and that logically fits at /participant/<pk>.

Although we don't link to it directly, the /participants/<pk> URL is still visitable:
- when you click on back when visiting any of the URLs nested under it,
  and the `return_url` query param is missing for any reason
- when you go directly to the URL

This change is a quick fix that preserves all the existing links,
but takes users to the right place if this URL is visited.

In the future we might consider improving the back links and
potentially migrating all these URLs to live under /mammograms, but
I don't want to get bogged down in this now.