add past mammograms to medical history page
This is not yet complete, and it only shows reported mammograms, not
mammograms taken from the screening record.

I'm just making it consistent with what's already shown on the view
appointment page, under participant details.

The copy is slightly misleading because the last confirmed mammogram
isn't shown yet.

The "Add another mammogram" button on the record medical
information page is a small button, and submitting the form
should redirect back to the record medical information page.

In the prototype, the form redirects back to the specific section,
and highlights it with a border. We don't have this yet, but we can
implement it consistently for all sections in another ticket.