ClinicSlot started_at_time added
ClinicSlots are created in the context of a Clinic which already contains the date, meaning the started_at datetime is redundant information, and may cause issues if the two are not lined up. This PR adds a started_at_time field which is a timefield to remove that issue. This is the first step to migrating to just time, the next PR will remove the started_at datetime field.