create summaryList variant for complex information
This uses the styles from the summary list but abandons the <dl> markup
so we can display more complex information with a nested structure.

Instead of using a definition list, organise the information under
regular headings.

If we don't include the heading structure, then the nested structure
isn't easily perceivable to screenreader users.

If we include headings within the `dd`, but keep the dl / dh / dt
structure, then the headings don't make sense when navigating by
heading using assistive tech.

Screenreaders don't always announce when leaving dls, so I've decided
not to try nesting them.

Another option is to use the reponsive table component, and get rid of
the nested grid layout. However, this would require altering the overall
positioning of the blocks of information, and in our case this is quite
deliberate: right is always displayed on the left, and left on the
right, with systemic information always displayed below.

I've tested this component in Safari / VoiceOver and Edge / NVDA.