create macros for breast cancer history item cards
These cards are displayed within the overall card for medical history,
and are grouped under a section heading of "Breast cancer history".

Since the design system card component is currently being updated,
I've left the change link as a standalone link for now. When
the new release is out, we'll move it into the card.