show breast cancer history cards on the medical info page
This introduces the *card content* for the innermost cards that display
breast cancer history items within the medical history card.

When there is more than one of these cards, they have card headings,
Item 1, Item 2 etc, which affects the overall heading hierarchy.
If there is only one, this heading is left out.

In a future commit, I'll align the card part with the latest card from
the design system, and move the change links into the card.
At that point we may want to also remove the padding around the card
content.

The prototype is using a custom card design, so I'll see which
customisations are still necessary after the new design system version
is out.